+++ src/printer.rs
@@ -381,7 +381,7 @@ impl<'a> Printer for InteractivePrinter<'a> {
         let line = if self.config.show_nonprintable {
             replace_nonprintable(line_buffer, self.config.tab_width)
         } else {
-            match self.content_type {
+            let line = match self.content_type {
                 Some(ContentType::BINARY) | None => {
                     return Ok(());
                 }
@@ -392,6 +392,14 @@ impl<'a> Printer for InteractivePrinter<'a> {
                     .decode(line_buffer, DecoderTrap::Replace)
                     .map_err(|_| "Invalid UTF-16BE")?,
                 _ => String::from_utf8_lossy(line_buffer).to_string(),
+            };
+
+            // Remove byte order mark from interactive output.
+            // Kept if show_nonprintable.
+            if line_number == 1 {
+                line.replace('\u{feff}', "")
+            } else {
+                line
             }
         };