+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Python syntax highlighting no longer suffers from abysmal performance in specific scenarios. See #1688 (@keith-hall)
 - First line not shown in diff context. See #1891 (@divagant-martian)
 - Do not ignore syntaxes that handle file names with a `*.conf` extension. See #1703 (@cbolgiano)
+- Strips byte order mark from output in interactive mode. See #1922 (@dag-h)
 
 ## Performance